+++ src/result_methods.jl
@@ -1,9 +1,17 @@
 """
-    GEEBRA_results(results::Union{NLsolve.SolverResults, Optim.MultivariateOptimizationResults, Optim.UnivariateOptimizationResults}, theta::Vector, data::Any,  data::Any, template::Union{objective_function_template, estimating_function_template}, br::Bool, has_objective::Bool, br_method::String)
+    MEstimation_results(results::Union{NLsolve.SolverResults, Optim.MultivariateOptimizationResults, Optim.UnivariateOptimizationResults},
+                        theta::Vector,
+                        data::Any,
+                        template::Union{objective_function_template, estimating_function_template},
+                        regularizer::Function,
+                        br::Bool,
+                        has_objective::Bool,
+                        has_regularizer::Bool,
+                        br_method::String)
 
 Composite type for the output of [`fit`](@ref) for an [`objective_function_template`](@ref) or an [`estimating_function_template`](@ref).
 """
-struct GEEBRA_results
+struct MEstimation_results
     results::Union{NLsolve.SolverResults, Optim.MultivariateOptimizationResults, Optim.UnivariateOptimizationResults}
     theta::Vector
     data::Any
@@ -16,11 +24,11 @@ struct GEEBRA_results
 end
 
 """
-    vcov(results::GEEBRA_results)
+    vcov(results::MEstimation_results)
 
 Compute an esitmate of the variance-covariance matrix of the `M`-estimator or its reduced-bias version from the output of [`fit`](@ref) for an [`objective_function_template`](@ref) or an [`estimating_function_template`](@ref).
 """
-function vcov(results::GEEBRA_results)
+function vcov(results::MEstimation_results)
     if (results.has_objective)
         obj_quantities(results.theta, results.data, results.template, false)[1]
     else
@@ -29,11 +37,11 @@ function vcov(results::GEEBRA_results)
 end
 
 """
-    tic(results::GEEBRA_results)
+    tic(results::MEstimation_results)
 
 Compute the Takeuchi Information Criterion at the `M`-estimator or its reduced-bias version from the output of [`fit`](@ref) for an [`objective_function_template`](@ref). `nothing` is returned if `results` is the output of [`fit`](@ref) for an [`estimating_function_template`](@ref).
 """
-function tic(results::GEEBRA_results)
+function tic(results::MEstimation_results)
     if (results.has_objective)
         obj = objective_function(results.theta, results.data, results.template, false)
         quants = obj_quantities(results.theta, results.data, results.template, true)
@@ -42,11 +50,11 @@ function tic(results::GEEBRA_results)
 end
 
 """
-    aic(results::GEEBRA_results)
+    aic(results::MEstimation_results)
 
 Compute the Akaike Information Criterion at the `M`-estimator or its reduced-bias version from the output of [`fit`](@ref) for an [`objective_function_template`](@ref). `nothing` is returned if `results` is the output of [`fit`](@ref) for an [`estimating_function_template`](@ref).
 """
-function aic(results::GEEBRA_results)
+function aic(results::MEstimation_results)
     if (results.has_objective)
         obj = objective_function(results.theta, results.data, results.template, false)
         p = length(results.theta)
@@ -55,20 +63,20 @@ function aic(results::GEEBRA_results)
 end
 
 """
-    coef(results::GEEBRA_results)
+    coef(results::MEstimation_results)
 
 Extract the `M`-estimates or their reduced-bias versions from the output of [`fit`](@ref) for an [`objective_function_template`](@ref) or an [`estimating_function_template`](@ref).
 """
-function coef(results::GEEBRA_results)
+function coef(results::MEstimation_results)
     results.theta
 end
 
 """
-    show(io::IO, results::GEEBRA_results; digits::Real = 4)
+    show(io::IO, results::MEstimation_results; digits::Real = 4)
 
-`show` method for `GEEBRA_results` objects. If `GEEBRA_results.has_object == true`, then the result of `aic(results)` and `tic(results)` are also printed.
+`show` method for `MEstimation_results` objects. If `MEstimation_results.has_object == true`, then the result of `aic(results)` and `tic(results)` are also printed.
 """
-function Base.show(io::IO, results::GEEBRA_results;
+function Base.show(io::IO, results::MEstimation_results;
                    digits::Real = 4)
     theta = results.theta
     p = length(theta)
@@ -127,20 +135,21 @@ function Base.show(io::IO, results::GEEBRA_results;
 end
 
 """
-    stderror(results::GEEBRA_results)
+    stderror(results::MEstimation_results)
 
 Compute esitmated standard errors for the `M`-estimator or its reduced-bias version from the output of [`fit`](@ref) for an [`objective_function_template`](@ref) or an [`estimating_function_template`](@ref).
 """
-function stderror(results::GEEBRA_results)
+function stderror(results::MEstimation_results)
     sqrt.(diag(vcov(results)))
 end
 
 """
-    coeftable(results::GEEBRA_results; level::Real=0.95)
+    coeftable(results::MEstimation_results; 
+              level::Real=0.95)
 
 Return a `StatsBase.CoefTable` for the `M`-estimator or its reduced-bias version from the output of [`fit`](@ref) for an [`objective_function_template`](@ref) or an [`estimating_function_template`](@ref). `level` can be used to set the level of the reported Wald-type confidence intervals (using quantiles of the standard normal distribution). 
 """
-function coeftable(results::GEEBRA_results; level::Real=0.95)
+function coeftable(results::MEstimation_results; level::Real=0.95)
     cc = coef(results)
     se = stderror(results)
     zz = cc ./ se