+++ NEWS.md
@@ -2,8 +2,11 @@
 
 + Added support for bias reduction of a subset of the parameters (only for `estimating_function_template`)
 + Added option for optimization in constrained parameters spaces (only for `objective_function_template`)
++ Added option to provide own regularizer function (either for objective or estimating functions)
++ Added slices for computing slices of objectives and estimating functions
 + The default output now reports whether the optimization algorithm converged or not
 + Documentation improvements and updates
++ Optimization improvements
 
 
 # GEEBRA 0.1.0